+++ examples/network-error.ts
@@ -8,7 +8,7 @@
  */
 
 import axios from 'axios';
-import createAuthRefreshInterceptor from '../src/index';
+import { createAuthRefresh } from '../src/index';
 import { createMockAdapter, MockState } from './_helpers/mock-adapter';
 import { assertEqual } from './_helpers/assert';
 
@@ -23,7 +23,7 @@ const instance = axios.create({
     headers: { Authorization: 'Bearer token-v1' },
 });
 
-createAuthRefreshInterceptor(
+createAuthRefresh(
     instance,
     async (failedRequest) => {
         state.refreshCount++;