+++ src/index.ts
@@ -28,7 +28,7 @@ export { AxiosAuthRefreshOptions, AxiosAuthRefreshRequestConfig } from './model'
  * @param {AxiosAuthRefreshOptions} options - options for the interceptor @see defaultOptions
  * @return {number} - interceptor id (in case you want to eject it manually)
  */
-export default function createAuthRefreshInterceptor(
+export function createAuthRefresh(
     instance: AxiosInstance,
     refreshAuthCall: (error: any) => Promise<any>,
     options: AxiosAuthRefreshOptions = {},
@@ -78,3 +78,7 @@ export default function createAuthRefreshInterceptor(
         },
     );
 }
+
+/** @deprecated Use createAuthRefresh instead */
+export const createAuthRefreshInterceptor = createAuthRefresh;
+export default createAuthRefresh;